feat: moved arc starknet wrapper type into seperate file
